reimplement observe.buyer.jpool.pool-kv

+++ sh/template/generate_xls_juben.sh
@@ -2,6 +2,7 @@
 
 # 仅用于根据模板(observe/buyer/template/xls.juben.properties)生成题材的剧本文件(即stk_daily/template/xxx.juben.properties) 
 # Usage: sh/template/generate_xls_juben.sh xls [template2] [--day ] 	-> 注意这里的template2 即支持指定名字,默认为xls
+# 注意: 会做replace的操作
 
 path=`pwd`
 export PYTHONPATH=$path:$PYTHONPATH
@@ -43,6 +44,9 @@ do
 		elif [ $now -eq 1 ]
 		then
 			template2=$1
+		elif [ $now -eq 2 ]
+		then
+			day=$1
 		fi
 		declare -i now=$now+1
 		;;
@@ -53,7 +57,6 @@ done
 cur_dir=/Users/wuxian/Desktop/STK-Experiment
 
 file1="/Users/wuxian/Desktop/stk_daily/$day/template/$xls.juben.properties"
-
 file2="$cur_dir/engine/observe/buyer/template/$template2.juben.properties"
 
 # 检验file1是否已经存在
@@ -68,6 +71,19 @@ then
 	echo 模板剧本文件: $file2 不存在,是否输入错误 比如改为 sh/template/generate_xls_juben.sh $xls xx_$template2  ?
 	exit 2
 fi
+
+path=`pwd`
+export PYTHONPATH=$path:$PYTHONPATH
+
+# 做一下校验replace kv逻辑
+check_replace=$(python engine/observe/buyer/template/juben_cli.py check_xls_by_replace_kv $xls $file2)
+
+if [[ $check_replace == "0" ]]
+then
+	echo python engine/observe/buyer/template/juben_cli.py check_xls_by_replace_kv 返回0,说明输入的xls:$xls 和对应文件中的replace kv不匹配
+	exit 2
+fi
+
 
 echo "复制文件 cp $file2 $file1"
 cp $file2 $file1
